combine all password related queries in password.js
